requirements.txt
transformers
torch
langchain
sentence-transformers
langchain_community
accelerate
pypdf
faiss-cpu
